chemical-equipment-visualizer/backend/requirements.txt
Django>=4.0
djangorestframework
django-filter
pandas
reportlab
django-cors-headers
whitenoise
